+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 ## [Unreleased]
 ### Added
 - Ported to 1.18.2
+- Added Smooth Night Vision
 
 ### Changed
 - Detectors no longer conduct redstone power